pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tradestation"
version = "0.2"
description = "An unofficial, sync/async SDK for Tradestation!"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  {name = "Graeme Holliday", email = "graeme@tastyware.dev"},
  {name = "Ethan Corgatelli", email = "ethan@tastyware.dev"},
]
dependencies = [
    "httpx>=0.27.2",
    "pydantic>=2.9.2",
    "pyjwt>=2.10.1",
]

[project.urls]
Homepage = "https://github.com/tastyware/tradestation"
Documentation = "https://tradestation.readthedocs.io/en/latest"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.7",
    "pyright>=1.1.391",
    "pytest-aio>=1.9.0",
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=3.0.2",
    "autodoc-pydantic>=2.2.0",
    "enum-tools[sphinx]>=0.12.0",
]

[tool.setuptools.package-data]
"tradestation" = ["py.typed"]

[tool.setuptools.packages.find]
where = ["tradestation"]
